+++ tests/testthat/helper-map.R
@@ -1,2 +1,5 @@
 
 named <- function(x) set_names(x, chr())
+
+# Until we can reexport from rlang
+vars <- quos