+++ NEWS.md
@@ -214,6 +214,9 @@ arguments.
 * `pluck()` no longer flattens lists of arguments. You can still do it
   manually with `!!!`. This change is for consistency with other
   dots-collecting functions of the tidyverse.
+  
+* `detect()` now has a .default argument to specify the value returned when 
+  nothing is detected (#622, @ColinFay)
 
 
 ## Life cycle